multishell - flesh out history provisions, including deletion.
Users are offered the option to delete a history entry for a shell
buffer when the buffer is killed.

Note that repair of a multishell bug, so tramp homedir syntax
(`/example.com:') is now recognized, seems to expose a tramp bug. Tramp
sometimes fails to open remote shells with the homedir syntax,
especially for remote+sudo. The failure is accompanied by a "Selecting
deleted buffer" error.