multishell - Remove notes about tramp remote+sudo+homedir+sudo prob
Apparently it's due to something in my local site configuration. The
problem happens reliably in emacs started with -q, but I can't get it to
happen if emacs was started with -Q.

Unfortunately, I don't get a traceback, even with tramp-verbose=6, when
the problem occurs, only "Selecting deleted buffer", and in *Messages*,
"byte-code: Selecting deleted buffer". Nothing more happens if
debug-on-error and debug-on-signal are set. This seems to suggest that
something is swallowing the signal, yet still returning to
top-level. Elusive.