multishell - method to conclude minibuffer interaction from multishell-list
Use throw-catch facility for non-local minibuffer exit. Originally
implemented for invoking and passing control to the minibuffer-list from
the minibuffer, it proves to be much more tidy than the non-local
minibuffer exit provisions in, eg, buff-menu.

For now, the most crucial binding in this context is the
mouse-click. The keyboard cases are also handled, but generally
switching the *Completions* buffer to the target one.